chores: rename src file, delete dead code
