+++ 00_HelloWorld/CMakeLists.txt
@@ -5,7 +5,7 @@ cmake_minimum_required(VERSION 3.10)
 find_package(PCL 1.13 REQUIRED COMPONENTS common io visualization)
 
 # Define the executable for this subproject
-add_executable(hello_pcl main.cpp)
+add_executable(hello_pcl hello.cpp)
 
 # Link the necessary libraries
 target_link_libraries(hello_pcl ${PCL_LIBRARIES})
